add rudimentary geom_mtt and mtt_calc
mtt_calc and geom_mtt are probably best used together. mtt_calc does all
the stuff that mtt_fit does in terms of calculation (subtraction of
background and normalizing to lowest condition concentration) but
without the fitting part. geom_mtt in theory should perform the same
fitting procedure at plot time. This is quite a bit easier to fiddle
with to make custom plots, so despite its crudeness, it's much easier to
work with.